2.0.x (#42)
* fix compilation on openbsd (#3814) (#3816)

the getentropy fallback is only possible on macOS, wrap it in a define
to remove it entirely on the BSDs.

* Backport release automation fixes (#3818)

* Refs #3646 -- Don't use shell globbing in our release script (#3809)

* fix wheel building with parameters (#3808)

* Refs #3646 -- invoke twine correctly (#3789)

* backport wheel builder fixes (#3817)

* update wheel builder script for static linking on linux (#3811)

We don't need to do an LD_LIBRARY_PATH when calling auditwheel because
we're now statically linking OpenSSL.

* Fixed #3801 -- don't create py33 wheels (#3802)

* Fixed #3798 -- use the correct stacklevel on this warning (#3803) (#3819)

* 2.0.1 changelog and version bump (#3820)

* don't re-export the openssl symbols from our shared object (#3825) (#3827)

When linking the static lib we want the symbols to not be exported. To
convince gcc to do this you must pass -Wl,--exclude-libs,ALL in CFLAGS
More info at https://stackoverflow.com/questions/2222162

* changelog and version bump for 2.0.2 (#3826)

* changelog and version bump for 2.0.2

* diff language

* Blacklist pytest 3.2.0 because it breaks us (#3835) (#3844)

* Blacklist pytest 3.2.0 because it breaks us (#3835)

See upstream: https://github.com/pytest-dev/pytest/issues/2644

* fix for latest flake8 (#3829)

* Fix weak linking of getentropy when compiling on older macOS (#3841) (#3842)

* Fix weak linking of getentropy when compiling on older macOS

We use weak linking in macOS to determine if the getentropy symbol is
available. However, to do that we need to have a declaration that states
the function is __attribute((weak_import)) at compile time. On macOS
10.12 this is provided in sys/random.h, but on older macOS the
declaration doesn't exist at all, so we need to forward declare it
ourselves.

* update a comment and a style nit

* 2.0.3 changelog and version bump (#3843)

Co-authored-by: Paul Kehrer <paul.l.kehrer@gmail.com>